Refactor scan I/O, device handling, CLI
Introduce ScanFactory and scan metadata, refactor scan/coarse_channel internals, and improve CLI tools and docs.

Key changes:
- Add ScanFactory (file_types/scan_factory.cpp/hpp) and scan metadata/data source headers to centralize file loading.
- Update CLI tools (bliss_find_hits, bliss_debug_demo, bliss_generate_channelizer_response, bliss_hits_to_dat) to use ScanFactory, add documentation/comments, improve argument handling and output behavior.
- Refactor coarse_channel to use metadata-backed constructors, support lazy loading, cache computed resources (hits, integrated drift planes), and defer/ensure device transfers (set_device/push_device and on-access transfers).
- Add helpers in core (extract_source_name_from_scans) and tighten validation/consistency checks across observation_target/cadence.
- Update CMake targets: link file_types into bliss_find_hits; remove file_types from bliss_core private link.
- Numerous CLI/flagging/noise-estimation/processing workflow clarifications and small behavior fixes (deprecated-argument warnings, output filename auto-detection, explicit device setting for scans).

These changes centralize file I/O, improve memory/device management for CPU/GPU workflows, and make the command-line tools more robust and better documented.

+++ bliss/preprocess/excise_dc.cpp
@@ -1,4 +1,3 @@
-
 #include <preprocess/excise_dc.hpp>
 
 #include <bland/ndarray.hpp>
@@ -9,14 +8,35 @@
 #include <fmt/ranges.h>
 
 #include <cmath>
+#include <cstdint> 
 
 using namespace bliss;
 
 coarse_channel bliss::excise_dc(coarse_channel cc) {
     auto data = cc.data();
-    auto dc_bin = data.size(1)/2;
-    auto dc_data = bland::slice(data, 1, dc_bin, dc_bin+1, 1);
-    dc_data = (bland::slice(data, 1, dc_bin-1, dc_bin, 1) + bland::slice(data, 1, dc_bin+1, dc_bin+2, 1))/2;
+    // Assuming the DC bin is exactly in the middle (FFT shift logic)
+    auto dc_bin = static_cast<int64_t>(data.size(1)/2);
+    
+    // Select the DC bin
+    auto dc_data = bland::slice(data, 
+                                static_cast<int64_t>(1), 
+                                dc_bin, 
+                                dc_bin + 1, 
+                                static_cast<int64_t>(1));
+    
+    // Replace it with the average of the left and right neighbors
+    // (left_neighbor + right_neighbor) / 2
+    dc_data = (bland::slice(data, 
+                            static_cast<int64_t>(1), 
+                            dc_bin - 1, 
+                            dc_bin, 
+                            static_cast<int64_t>(1)) 
+             + bland::slice(data, 
+                            static_cast<int64_t>(1), 
+                            dc_bin + 1, 
+                            dc_bin + 2, 
+                            static_cast<int64_t>(1))) / 2;
+                            
     cc.set_data(data);
     return cc;
 }
